# api-spec.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: >-
      https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  /keywords:
    post:
      tags:
        - 'UCM6: Keyword management'
        - Keywords
      summary: Create a new system-wide keyword
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '201':
          description: Keyword created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  keywordId:
                    type: integer
                    description: The ID of the keyword that has just been created
                    example: 600
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Keyword'
  /keywords/search:
    get:
      tags:
        - 'UCM6: Keyword management'
        - Keywords
      summary: Search for keywords by partial (or complete) name
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: query
          name: searchQuery
          schema:
            type: string
          required: true
          description: >-
            Criteria to search keywords for, all partial and complete matches
            are returned
      responses:
        '200':
          description: 'Request successful, a (possibly empty) list of keywords is returned'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Keyword'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  '/keywords/{id}':
    delete:
      tags:
        - 'UCM6: Keyword management'
        - Keywords
      summary: Delete a keyword as an admin
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of keyword to delete
          example: 600
      responses:
        '200':
          description: Keyword successfully deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned if a non-GAA user attempts to delete any keyword.'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: >-
        There was some error with the data supplied by the user, appropriate
        error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
  schemas:
    Keyword:
      type: object
      description: Represents a single 'keyword' within the system - similar to a hashtag.
      properties:
        id:
          type: integer
          example: 600
          readOnly: true
          description: 'Unique ID, autogenerated by the system'
        name:
          type: string
          description: The text value of the keyword
          example: Vehicle
        created:
          type: string
          format: date-time
          example: '2021-07-15T05:10:00Z'
          readOnly: true
          description: Date and time when the keyword was created
...